in place replacement
