+++ src/views/MarketDetails/CurrentMarketDetails.tsx
@@ -1,9 +1,11 @@
+import { selectCurrentMarketInfo } from '@/abacus-ts/selectors/markets';
+import { IndexerPerpetualMarketType } from '@/types/indexer/indexerApiGen';
 import BigNumber from 'bignumber.js';
 import { shallowEqual } from 'react-redux';
 
-import { PerpetualMarketType } from '@/constants/abacus';
 import { STRING_KEYS } from '@/constants/localization';
 
+import { useMetadataServiceAssetFromId } from '@/hooks/useMetadataService';
 import { useStringGetter } from '@/hooks/useStringGetter';
 
 import { DetailsItem } from '@/components/Details';
@@ -11,56 +13,49 @@ import { DiffOutput } from '@/components/DiffOutput';
 import { Output, OutputType } from '@/components/Output';
 
 import { useAppSelector } from '@/state/appTypes';
-import { getCurrentMarketAssetData } from '@/state/assetsSelectors';
-import { getCurrentMarketData } from '@/state/perpetualsSelectors';
 
-import { getDisplayableAssetFromBaseAsset } from '@/lib/assetUtils';
 import { BIG_NUMBERS } from '@/lib/numbers';
+import { orEmptyObj } from '@/lib/typeUtils';
 
 import { MarketDetails } from './MarketDetails';
 
 export const CurrentMarketDetails = () => {
   const stringGetter = useStringGetter();
-  const { configs, displayId } = useAppSelector(getCurrentMarketData, shallowEqual) ?? {};
-  const { id, name, resources } = useAppSelector(getCurrentMarketAssetData, shallowEqual) ?? {};
-
-  if (!configs) return null;
+  const currentMarketData = useAppSelector(selectCurrentMarketInfo, shallowEqual);
+  const asset = useMetadataServiceAssetFromId(currentMarketData?.ticker);
 
   const {
-    tickSize,
-    stepSize,
-    initialMarginFraction,
+    displayableAsset,
+    displayableTicker,
     effectiveInitialMarginFraction,
+    initialMarginFraction,
     maintenanceMarginFraction,
-    minOrderSize,
-    perpetualMarketType,
-    stepSizeDecimals,
+    marketType,
+    tickSize,
     tickSizeDecimals,
-  } = configs;
+    stepSize,
+    stepSizeDecimals,
+  } = orEmptyObj(currentMarketData);
 
-  const {
-    coinMarketCapsLink,
-    primaryDescriptionKey,
-    secondaryDescriptionKey,
-    websiteLink,
-    whitepaperLink,
-  } = resources ?? {};
+  const { id: assetId, logo, name, urls } = orEmptyObj(asset);
+  const { cmc, website, technicalDoc } = orEmptyObj(urls);
 
   const preferEIMF = Boolean(
-    effectiveInitialMarginFraction && initialMarginFraction !== effectiveInitialMarginFraction
+    effectiveInitialMarginFraction &&
+      initialMarginFraction !== effectiveInitialMarginFraction.toString()
   );
 
   const items = [
     {
       key: 'ticker',
       label: stringGetter({ key: STRING_KEYS.TICKER }),
-      value: displayId,
+      value: displayableTicker,
     },
     {
       key: 'market-type',
       label: stringGetter({ key: STRING_KEYS.TYPE }),
       value:
-        perpetualMarketType === PerpetualMarketType.CROSS
+        marketType === IndexerPerpetualMarketType.CROSS
           ? stringGetter({ key: STRING_KEYS.CROSS })
           : stringGetter({ key: STRING_KEYS.ISOLATED }),
     },
@@ -86,7 +81,7 @@ export const CurrentMarketDetails = () => {
           useGrouping
           value={stepSize}
           type={OutputType.Asset}
-          tag={getDisplayableAssetFromBaseAsset(id)}
+          tag={displayableAsset}
           fractionDigits={stepSizeDecimals}
         />
       ),
@@ -97,9 +92,9 @@ export const CurrentMarketDetails = () => {
       value: (
         <Output
           useGrouping
-          value={minOrderSize}
+          value={stepSize}
           type={OutputType.Asset}
-          tag={getDisplayableAssetFromBaseAsset(id)}
+          tag={displayableAsset}
           fractionDigits={stepSizeDecimals}
         />
       ),
@@ -149,11 +144,11 @@ export const CurrentMarketDetails = () => {
   return (
     <MarketDetails
       assetName={name}
-      assetIcon={{ symbol: id, logoUrl: resources?.imageUrl }}
+      assetIcon={{ symbol: assetId, logoUrl: logo }}
       marketDetailItems={items}
-      primaryDescription={stringGetter({ key: `APP.${primaryDescriptionKey}` })}
-      secondaryDescription={stringGetter({ key: `APP.${secondaryDescriptionKey}` })}
-      urls={{ technicalDoc: whitepaperLink, website: websiteLink, cmc: coinMarketCapsLink }}
+      primaryDescription={stringGetter({ key: `APP.__ASSETS.${assetId}.PRIMARY` })}
+      secondaryDescription={stringGetter({ key: `APP.__ASSETS.${assetId}.SECONDARY` })}
+      urls={{ technicalDoc, website, cmc }}
     />
   );
 };
